requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
Pillow==10.4.0
aiohttp==3.9.5
PyNaCl==1.5.0
yt-dlp
humanize==4.9.0
aiofiles==23.2.1
psutil==5.9.8
google-generativeai==0.5.4
matplotlib==3.8.4
aiosqlite==0.19.0
requests==2.32.3
cryptography==42.0.8
itsdangerous==2.2.0